Move identical code to AbstarctCassandraJoin trait
